requirements.txt
arrow==1.2.1
asgiref==3.5.0
binaryornot==0.4.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cookiecutter==1.7.3
Django==4.0.3
django-browser-reload==1.3.0
django-environ==0.8.1
django-phonenumber-field==6.1.0
django-tailwind==3.0.1
idna==3.3
Jinja2==3.1.1
jinja2-time==0.2.0
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.2
phonenumbers==8.12.46
poyo==0.5.0
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2022.1
urllib3==1.26.9
